cherry-pick new opengl features keeping bluetooth classes

+++ app/src/main/cpp/native-lib.cpp
@@ -16,7 +16,7 @@ extern "C" {
     }
 
     JNIEXPORT void JNICALL Java_com_example_android_sonar_SonarJNI_draw(JNIEnv* env, jclass type,
-                                                                        jint distance, jint angle)
+                                                                        jfloat distance, jint angle)
     {
         Engine::draw_frame(distance, angle);
     }
