+++ app/src/main/java/com/example/android/sonar/SonarJNI.java
@@ -7,5 +7,5 @@ public class SonarJNI {
 
     public static native void surface_created();
     public static native void surface_changed(int width, int height);
-    public static native void draw(int distance, int angle);
+    public static native void draw(float distance, int angle);
 }
